requirements.txt
pulumi==3.46.0
pulumi-gcp==6.41.0
pulumi-google-native==0.27.0
pulumi-aws==5.9.0
pulumi-eks==0.40.0
pulumi-kubernetes==3.18.1
requests==2.27.1
paramiko==2.11.0
ruamel.yaml==0.17.21
GitPython==3.1.29
boto3==1.24.57
botocore==1.27.57